Properly handle loading order and pair names again
Seems like using bank ids from the swdl header was a flop.  Turns out there's a lot of collision.
So I reverted to the old way of just using the file name, and I also stored the order in which each file name or "pair name" was loaded. That way a swdl + smdl pair loaded independently in any order will always have the matching name and number when exported. It makes things much easier to handle.